refactor(All): domain 및 controller 패키지 하위로 이동
