+++ src/main/java/lotto/controller/LottoGame.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.controller;
 
 public class LottoGame {
