+++ src/main/java/lotto/domain/LottoMachine.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import static lotto.global.ErrorMessage.*;
 import static lotto.global.LottoConstant.*;
